% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_vegan.R
\name{register_vegan}
\alias{register_vegan}
\alias{isomap}
\alias{monoMDS}
\alias{metaMDS}
\title{Register Seriation Methods from Package vegan}
\usage{
register_vegan()
}
\value{
Nothing.
}
\description{
Register the \code{"isomap"}, \code{"monoMDS"}, and \code{"metaMDS"} seriation methods
for dissimilarity matrices. These methods use the corresponding functions
from package \pkg{vegan}.
}
\details{
\strong{Note:} Package \pkg{vegan} needs to be installed.
}
\examples{
\dontrun{
register_vegan()
list_seriation_methods("dist")
}
}
\seealso{
Other seriation:
\code{\link[=register_DendSer]{register_DendSer()}},
\code{\link[=register_GA]{register_GA()}},
\code{\link[=register_optics]{register_optics()}},
\code{\link[=register_smacof]{register_smacof()}},
\code{\link[=register_tsne]{register_tsne()}},
\code{\link[=register_umap]{register_umap()}},
\code{\link{registry_for_seriation_methods}},
\code{\link[=seriate]{seriate()}},
\code{\link[=seriate_best]{seriate_best()}}
}
\concept{seriation}
