% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chameleon.R
\docType{data}
\name{Chameleon}
\alias{Chameleon}
\alias{chameleon}
\alias{chameleon_ds4}
\alias{chameleon_ds5}
\alias{chameleon_ds7}
\alias{chameleon_ds8}
\title{2D Data Sets used for the CHAMELEON Clustering Algorithm}
\format{
\code{chameleon_ds4}: The format is a 8,000 x 2 data.frame.

\code{chameleon_ds5}: The format is a 8,000 x 2 data.frame.

\code{chameleon_ds7}: The format is a 10,000 x 2 data.frame.

\code{chameleon_ds8}: The format is a 8,000 x 2 data.frame.
}
\description{
Several 2D data sets created to evaluate the CHAMELEON clustering algorithm in
the paper by Karypis et al (1999).
}
\examples{
data(Chameleon)

plot(chameleon_ds4, cex = .1)
plot(chameleon_ds5, cex = .1)
plot(chameleon_ds7, cex = .1)
plot(chameleon_ds8, cex = .1)
}
\references{
Karypis, G., EH. Han, V. Kumar (1999): CHAMELEON: A Hierarchical
Clustering Algorithm Using Dynamic Modeling, \emph{IEEE Computer,}
\strong{32}(8): 68--75.
\doi{10.1109/2.781637}
}
\seealso{
Other data:
\code{\link{Irish}},
\code{\link{Munsingen}},
\code{\link{Psych24}},
\code{\link{SupremeCourt}},
\code{\link{Townships}},
\code{\link{Wood}},
\code{\link{Zoo}},
\code{\link[=create_lines_data]{create_lines_data()}},
\code{\link[=is.robinson]{is.robinson()}}
}
\concept{data}
\keyword{datasets}
