% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irish.R
\docType{data}
\name{Irish}
\alias{Irish}
\title{Irish Referendum Data Set}
\format{
The format is a 41 x 9 matrix. Two values are missing.
}
\source{
The data was kindly provided by Guenter Sawitzki.
}
\description{
A data matrix containing the results of 8 referenda for 41 Irish communities
used in Falguerolles et al (1997).
}
\details{
Column 6 contains the size of the Electorate in 1992.
}
\examples{
data(Irish)
}
\references{
de Falguerolles, A., Friedrich, F., Sawitzki, G. (1997) A
Tribute to J. Bertin's Graphical Data Analysis. In: \emph{Proceedings of the
SoftStat '97 (Advances in Statistical Software 6),} 11--20.
}
\seealso{
Other data:
\code{\link{Chameleon}},
\code{\link{Munsingen}},
\code{\link{Psych24}},
\code{\link{SupremeCourt}},
\code{\link{Townships}},
\code{\link{Wood}},
\code{\link{Zoo}},
\code{\link[=create_lines_data]{create_lines_data()}},
\code{\link[=is.robinson]{is.robinson()}}
}
\concept{data}
\keyword{datasets}
