\name{Irish}
\alias{Irish}
\docType{data}
\title{Irish Referendum Data Set}
\description{
A data matrix containing the results of 8 referenda for 41 Irish communities 
 used in Falguerolles et al (1997).
}
\usage{data(Irish)}
\format{
  The format is a 41 x 9 matrix. Two values are missing.
}
\details{
Column 6 contains the size of the Electorate in 1992.
}
\source{
The data was kindly provided by Guenter Sawitzki.
}
\references{
de Falguerolles, A., Friedrich, F., Sawitzki, G. (1997):
A Tribute to J. Bertin's Graphical Data Analysis.
In: Proceedings of the SoftStat '97 (Advances in Statistical Software 6),
11--20.
}
\examples{
data(Irish)
}
\keyword{datasets}
