% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lle.R
\name{lle}
\alias{lle}
\alias{LLE}
\alias{tries}
\alias{to}
\alias{find}
\alias{a}
\alias{lower-dimensional}
\alias{projection}
\alias{which}
\alias{preserves}
\alias{distances}
\alias{within}
\alias{local}
\alias{neighborhoods.}
\alias{This}
\alias{is}
\alias{done}
\alias{by}
\alias{(1)}
\alias{for}
\alias{each}
\alias{object}
\alias{the}
\alias{k}
\alias{nearest}
\alias{neighbors,}
\alias{(2)}
\alias{construct}
\alias{weight}
\alias{matrix}
\alias{represents}
\alias{point}
\alias{as}
\alias{linear}
\alias{combination}
\alias{of}
\alias{its}
\alias{neighborhood,}
\alias{and}
\alias{perform}
\alias{partial}
\alias{eigenvalue}
\alias{decomposition}
\alias{embedding.}
\alias{The}
\alias{`reg`}
\alias{parameter}
\alias{allows}
\alias{decision}
\alias{between}
\alias{different}
\alias{regularization}
\alias{methods.}
\alias{As}
\alias{one}
\alias{step}
\alias{algorithm,}
\alias{inverse}
\alias{Gram-matrix}
\alias{\eqn{G\in}
\alias{R^{kxk}}}
\alias{has}
\alias{be}
\alias{calculated.}
\alias{rank}
\alias{\eqn{G}}
\alias{equals}
\alias{\eqn{m}}
\alias{mostly}
\alias{smaller}
\alias{than}
\alias{\eqn{k}}
\alias{-}
\alias{this}
\alias{why}
\alias{\eqn{G^{(i)}+r\cdot}
\alias{I}}
\alias{should}
\alias{performed.}
\alias{calculation}
\alias{\eqn{r}}
\alias{can}
\alias{using}
\alias{methods:}
\alias{`reg}
\alias{=}
\alias{1`:}
\alias{standardized}
\alias{sum}
\alias{eigenvalues}
\alias{(Roweis}
\alias{Saul;}
\alias{2000)}
\alias{2`}
\alias{(default):}
\alias{trace}
\alias{divided}
\alias{(Grilli,}
\alias{2007)}
\alias{3`:}
\alias{constant}
\alias{value}
\alias{3*10e-3}
\title{Locally Linear Embedding (LLE)}
\usage{
lle(x, m, k, reg = 2)
}
\arguments{
\item{x}{a matrix.}

\item{m}{dimensions of the desired embedding.}

\item{k}{number of neighbors.}

\item{reg}{regularization method. 1, 2 and 3, by default 2. See details.}
}
\value{
a matrix of vector with the embedding.
}
\description{
Performs the non linear dimensionality reduction method locally linear embedding
proposed in Roweis and Saul (2000).
}
\examples{
data(iris)
x <- iris[, -5]

# project iris on 2 dimensions
conf <- lle(x, m = 2, k = 30)
conf

plot(conf, col = iris[, 5])

# project iris onto a single dimension
conf <- lle(x, m = 1, k = 30)
conf

plot_config(conf, col = iris[, 5], labels = FALSE)
}
\references{
Roweis, Sam T. and Saul, Lawrence K. (2000), Nonlinear Dimensionality
Reduction by Locally Linear Embedding,
\emph{Science,} \strong{290}(5500), 2323--2326. \doi{10.1126/science.290.5500.2323}

Grilli, Elisa (2007) Automated Local Linear Embedding with an application
to microarray data, Dissertation thesis, University of Bologna.
\doi{10.6092/unibo/amsdottorato/380}
}
\author{
Michael Hahsler (based on code by Holger Diedrich and Markus Abel)
}
\keyword{cluster}
\keyword{manip}
