% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_smacof.R
\name{register_smacof}
\alias{register_smacof}
\alias{registersmacof}
\alias{smacof}
\title{Register Seriation Methods from Package smacof}
\usage{
register_smacof()
}
\value{
Nothing.
}
\description{
Registers methods for \code{\link[=seriate]{seriate()}} based on multidemensional scaling using
stress majorization implemented in package smacof (de Leeuw & Mair, 2009).
}
\details{
\strong{Rewite}
Registers the method \code{"DendSer"} for \code{\link[=seriate]{seriate()}}. DendSer is a fast
heuristic for reordering dendrograms developed by Earle and Hurley (2015)
able to use different criteria.

\code{control} for \code{seriate} with
method \code{"DendSer"} accepts the following parameters:
\itemize{
\item "h" or "method" A dendrogram or a method for hierarchical clustering
(see \code{hclust}). Default: complete-link.
\item "criterion" A seriation criterion to optimize (see
\code{list_criterion_methods("dist")}). Default: \code{"BAR"} (Banded
anti-Robinson from with 20\\% band width).}
\item "verbose" a logical; print progress information?
\item "DendSer_args" additional arguments for \code{DendSer}.
}

Note: Package \pkg{smacof} needs to be installed.
}
\examples{
\dontrun{
register_smacof()

get_seriation_method("dist", "nMDS")

d <- dist(random.robinson(20, pre = TRUE))

## use Banded AR form with default clustering (complete-link)
o <- seriate(d, "nMDS", verbose = TRUE)
pimage(d, o)

# recalculate stress for the order
MDS_stress(d, o)
}
}
\references{
Jan de Leeuw, Patrick Mair (2009). Multidimensional Scaling Using Majorization: SMACOF in R.
\emph{Journal of Statistical Software, 31(3),} 1-30. \doi{10.18637/jss.v031.i03}
}
\seealso{
Other seriation: 
\code{\link{register_DendSer}()},
\code{\link{register_GA}()},
\code{\link{register_optics}()},
\code{\link{register_tsne}()},
\code{\link{register_umap}()},
\code{\link{register_vae}()},
\code{\link{registry_for_seriaiton_methods}},
\code{\link{seriate_best}()},
\code{\link{seriate}()}
}
\concept{seriation}
\keyword{cluster}
\keyword{optimize}
