% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_tsne.R
\name{register_tsne}
\alias{register_tsne}
\alias{tsne}
\alias{tSNE}
\title{Register Seriation Based on 1D t-SNE}
\usage{
register_tsne()
}
\value{
Nothing.
}
\description{
Use t-distributed stochastic neighbor embedding (t-SNE) for \code{\link[=seriate]{seriate()}}.
}
\details{
Registers the method \code{"tsne"} for \code{\link[=seriate]{seriate()}}. This method applies
1D t-SNE to a data matrix or a distance matrix and extracts the order
from the 1D embedding. To speed up the process, an initial embedding is
created using 1D multi-dimensional scaling (MDS) or principal
comonents analysis (PCA) which is improved by t-SNE.

The \code{control} parameter \code{"mds"} or \code{"pca"} controls if MDS (for distances)
or PCA (for data matrices) is used to create an
initial embedding. See \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} to learn about the other
available \code{control} parameters.

Perplexity is automatically set as the minimum between 30 and the number of
observations. It can be also specified using the control parameter
\code{"preplexity"}.

\bold{Note:} Package \pkg{Rtsne} needs to be installed.
}
\examples{

\dontrun{
register_tsne()

# distances
get_seriation_method("dist", "tsne")

data(SupremeCourt)
d <- as.dist(SupremeCourt)

o <- seriate(d, method = "tsne")
pimage(d, o)

# look at the returned embedding and plot it
attr(o[[1]], "configuration")
plot_config(o)

## matrix
get_seriation_method("matrix", "tsne")

data("Zoo")
Zoo[,"legs"] <- (Zoo[,"legs"] > 0)
x <- as.matrix(Zoo[,-17])
label <- rownames(Zoo)
class <- Zoo$class

# tsne does not allow duplicates
x <- x[!duplicated(x), , drop = FALSE]

o <- seriate(x, method = "tsne")
pimage(x, o, prop = FALSE)

plot_config(o[[1]], col = class)
}

}
\references{
van der Maaten, L.J.P. & Hinton, G.E., 2008. Visualizing
High-Dimensional Data Using t-SNE. Journal of Machine Learning Research, 9,
pp.2579-2605.
}
\seealso{
\code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} in \pkg{Rtsne}.

Other seriation: 
\code{\link{register_DendSer}()},
\code{\link{register_GA}()},
\code{\link{register_optics}()},
\code{\link{register_smacof}()},
\code{\link{register_umap}()},
\code{\link{register_vae}()},
\code{\link{registry_for_seriaiton_methods}},
\code{\link{seriate_best}()},
\code{\link{seriate}()}
}
\concept{seriation}
\keyword{cluster}
\keyword{optimize}
