% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_registry_seriate.R
\name{registry_for_seriation_methods}
\alias{registry_for_seriation_methods}
\alias{registry_seriate}
\alias{list_seriation_methods}
\alias{get_seriation_method}
\alias{set_seriation_method}
\alias{print.seriation_method}
\title{Registry for Seriation Methods}
\usage{
registry_seriate

list_seriation_methods(kind, names_only = TRUE)

get_seriation_method(kind, name)

set_seriation_method(
  kind,
  name,
  definition,
  description = NULL,
  control = list(),
  randomized = FALSE,
  optimizes = NA_character_,
  verbose = FALSE,
  ...
)

\method{print}{seriation_method}(x, ...)
}
\arguments{
\item{kind}{the data type the method works on. For example, \code{"dist"},
\code{"matrix"} or \code{"array"}. If missing, then methods for any type are
shown.}

\item{names_only}{logical; return only the method name. \code{FALSE} returns
also the method descriptions.}

\item{name}{the name for the method used to refer to the method in
\code{\link[=seriate]{seriate()}}.}

\item{definition}{a function containing the method's code.}

\item{description}{a description of the method. For example, a long name.}

\item{control}{a list with control arguments and default values.}

\item{randomized}{logical; does the algorithm use randomization and re-running
the algorithm several times will lead to different results (see: \code{\link[=seriate_rep]{seriate_rep()}}).}

\item{optimizes}{what criterion does the algorithm try to optimize
(see: \code{\link[=list_criterion_methods]{list_criterion_methods()}}).}

\item{verbose}{logical; print a message when a new method is registered.}

\item{...}{further information that is stored for the method in the
registry.}

\item{x}{an object of class  "seriation_method" to be printed.}
}
\value{
\itemize{
\item \code{list_seriation_method()} result is a vector of character strings with the
names of the methods. These names are used for methods in \code{seriate()}.
\item \code{get_seriation_method()} returns a given method in form of an object of class
\code{"seriation_method"}.
}
}
\description{
A registry to manage methods used by \code{\link[=seriate]{seriate()}}.
}
\details{
The functions below are convenience function for the registry
\code{registry_seriate}.

\code{list_seriation_method()} lists all available methods for a given data
type (\code{kind}) (e.g., "dist", "matrix").
The result is a vector of character strings with the
method names that can be used in function \code{seriate()}.
If \code{kind} is missing, then a list of
methods is returned.

\code{get_seriation_method()} returns detailed information for a given method in
form of an object of class \code{"seriation_method"}.
The information includes a description, parameters and the
implementing function.

With \code{set_seriation_method()} new seriation methods can be added by the
user. The implementing function (\code{definition}) needs to have the formal
arguments \code{x, control} and, for arrays and matrices \code{margin},
where \code{x} is the data object and
\code{control} contains a list with additional information for the method
passed on from \code{seriate()}, and \code{margin} is a vector specifying
what dimensions should be seriated.
The implementation has to return a list of
objects which can be coerced into \code{ser_permutation_vector} objects
(e.g., integer vectors). The elements in the list have to be in
corresponding order to the dimensions of \code{x}.
}
\examples{
# Registry
registry_seriate

# List all seriation methods by type
list_seriation_methods()

# List methods for matrix seriation
list_seriation_methods("matrix")

get_seriation_method(name = "BEA")

# Example for defining a new seriation method (reverse identity function for matrix)

# 1. Create the seriation method: Reverse the row order
#    (NA means no seriation is applied to columns)
seriation_method_reverse_rows <- function(x, control = NULL, margin = c(1, 2)) {
    list(rev(seq(nrow(x))), NA)[margin]
}

# 2. Register new method
set_seriation_method("matrix", "Reverse_rows", seriation_method_reverse_rows,
    description = "Reverse identity order", control = list())

list_seriation_methods("matrix")
get_seriation_method("matrix", "reverse_rows")

# 3. Use the new seriation methods
seriate(matrix(1:12, ncol = 3), "reverse_rows")
}
\seealso{
This registry uses \link[registry:registry]{registry::registry}.

Other seriation:
\code{\link[=register_DendSer]{register_DendSer()}},
\code{\link[=register_GA]{register_GA()}},
\code{\link[=register_optics]{register_optics()}},
\code{\link[=register_smacof]{register_smacof()}},
\code{\link[=register_tsne]{register_tsne()}},
\code{\link[=register_umap]{register_umap()}},
\code{\link[=seriate]{seriate()}},
\code{\link[=seriate_best]{seriate_best()}}
}
\author{
Michael Hahsler
}
\concept{seriation}
\keyword{misc}
