% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriate_best.R
\name{seriate_best}
\alias{seriate_best}
\alias{seriate_rep}
\title{Best Seriation}
\usage{
seriate_best(
  x,
  methods = NULL,
  criterion = NULL,
  verbose = FALSE,
  fast_only = TRUE,
  ...
)

seriate_rep(x, method, criterion = NULL, rep = 10L, verbose = TRUE, ...)
}
\arguments{
\item{x}{the data.}

\item{methods}{a vector of character string with the name of the seriation
methods to try.}

\item{criterion}{a character string with the \link{criterion} to optimize.}

\item{verbose}{logical; show progress and results for different methods}

\item{fast_only}{logical; exclude slow default seriation methods.}

\item{...}{further arguments are passed on (e.g., as \code{control})}

\item{method}{a character string with the name of the seriation method
(default: varies by data type).}

\item{rep}{number of times to repeat the randomized seriation algorithm.}
}
\value{
Returns an object of class \link{ser_permutation}.
}
\description{
Functions to automatically try different seriation methods or
rerun randomized methods several times to find the best order
given a criterion measure.
}
\examples{
# prepare some datasets (two distance matrices and a data matrix)
data(SupremeCourt)
d_supreme <- as.dist(SupremeCourt)
d_robinson <- as.dist(random.robinson(50, noise = .2, pre = TRUE))
m_iris <- as.matrix(iris[ sample(seq(nrow(iris))),-5])

# find best seriation order
o <- seriate_best(d_supreme, verbose = TRUE)
pimage(d_supreme, o)

o <- seriate_best(d_robinson, verbose = TRUE)
pimage(d_robinson, o)

o <- seriate_best(m_iris, verbose = TRUE)
pimage(m_iris, o, prop = FALSE)

# run randomized algorithms several times. Repetition information
# is returned as attributes
o <- seriate_rep(d_supreme, "QAP_2SUM", rep = 10, verbose = TRUE)

attr(o, "criterion")
hist(attr(o, "criterion_distribution"))
pimage(d_supreme, o)

o <- seriate_rep(m_iris, "BEA_TSP", rep = 10, verbose = TRUE)

attr(o, "criterion")
hist(attr(o, "criterion_distribution"))
pimage(m_iris, o, prop = FALSE)
}
\seealso{
Other seriation: 
\code{\link{register_DendSer}()},
\code{\link{register_GA}()},
\code{\link{register_optics}()},
\code{\link{register_smacof}()},
\code{\link{register_tsne}()},
\code{\link{register_umap}()},
\code{\link{register_vae}()},
\code{\link{registry_for_seriaiton_methods}},
\code{\link{seriate}()}
}
\author{
Michael Hahsler
}
\concept{seriation}
\keyword{cluster}
\keyword{optimize}
