% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_seriation-package.R
\docType{package}
\name{seriation-package}
\alias{seriation-package}
\title{seriation: Infrastructure for Ordering Objects Using Seriation}
\description{
Infrastructure for ordering objects with an implementation of several seriation/sequencing/ordination techniques to reorder matrices, dissimilarity matrices, and dendrograms. Also provides (optimally) reordered heatmaps, color images and clustering visualizations like dissimilarity plots, and visual assessment of cluster tendency plots (VAT and iVAT).
}
\section{Key functions}{

\itemize{
\item Seriation: \code{\link[=seriate]{seriate()}}, \code{\link[=criterion]{criterion()}}, \code{\link[=get_order]{get_order()}}, \code{\link[=permute]{permute()}}
\item Visualization: \code{\link[=pimage]{pimage()}}, \code{\link[=bertinplot]{bertinplot()}}, \code{\link[=hmap]{hmap()}}, \code{\link[=dissplot]{dissplot()}}, \code{\link[=VAT]{VAT()}}
}
}

\references{
Michael Hahsler, Kurt Hornik, and Christian Buchta. Getting things in order: An introduction to the R package seriation. Journal of Statistical Software, 25(3):1--34, March 2008. \doi{10.18637/jss.v025.i03}
}
\author{
Michael Hahsler
}
